+++ src/components/features/life/solo-life-card.tsx
